move simpleblock to blocks package
